POST forecast (#5)
* Persist forecast data

* Some fixes

* Fix WorkflowEditor

* Rename UNITS_PER_HOUR

* Plural enum values

* Remove workflow from request body

* Add EXPEDITION process name

* Avoid hibernate logs at stage

* CR - code refactor

* CR - Remove unnecessary getUtcDate()

* CR - Remove unnecessary FetchMode

* Stop ignoring entity classes

* Add logs (#6)

* Round polyvalent productivity

* Check

* Fix gradlew checks

* Batch inserts

+++ src/main/java/com/mercadolibre/planning/model/api/domain/entity/current/CurrentHeadcountProductivity.java
@@ -37,7 +37,7 @@ public class CurrentHeadcountProductivity {
 
     private ProcessName processName;
 
-    private long productivity;
+    private double productivity;
 
     private MetricUnit productivityMetricUnit;
