+++ src/main/java/com/mercadolibre/planning/model/api/domain/entity/forecast/HeadcountDistribution.java
@@ -34,7 +34,7 @@ public class HeadcountDistribution {
     @Enumerated(EnumType.STRING)
     private ProcessName processName;
 
-    private long quantity;
+    private double quantity;
 
     @Enumerated(EnumType.STRING)
     private MetricUnit quantityMetricUnit;
